refactor: renaming and handling exports
